+++ .env.example
@@ -1,8 +1,3 @@
-OWNER_PRIVATE_KEY=
-ADDRESS_1_PRIVATE_KEY=
-ADDRESS_2_PRIVATE_KEY=
-ADDRESS_3_PRIVATE_KEY=
-FEEHANDLER_PRIVATE_KEY=
 POLYGON_MUMBAI_RPC_PROVIDER=
 POLYGON_RPC_PROVIDER=
 POLYGONSCAN_API_KEY=
